% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.lme.R
\name{.onAttach}
\alias{.onAttach}
\title{Power and Sample Size for Linear Mixed Effect Models Package  \cr}
\usage{
.onAttach(libname, pkgname)
}
\description{
Consider the following model: \cr \cr
      Y=XB+Zu+e, u~N(0,D), e~N(0,R) \cr
      Yi~N(XBi,Vi), Vi=Zi*D*Zi'+Ri, \cr \cr
          for each independent observation i \cr \cr \cr
estimate of fixed effect coefficient B, denoted by b: \cr \cr
      b=inv(sum(Xi'*inv(Vi)*Xi))*(sum(Xi'*inv(Vi)*Yi)) \cr \cr \cr
variance of b: \cr \cr
      var(b)=Vb/n=inv(sum(Xi'*inv(Vi)*Xi)) \cr \cr
          where Vb=inv(Xi'*inv(Vi)*Xi) \cr \cr \cr
distribution of any linear combinations L of b is given by: \cr \cr
      Lb~N(mu,Sigma/n) \cr \cr
            where mu = LB, Sigma = L*Vb*L' \cr \cr \cr
Interested parameters/linear combinations LB from more than one
independent populations can be aggregrate togeter by
appending mu vertically and Sigma/n diagonally \cr \cr
Consider Lb~N(MU,SIGMA) as the aggregrated estimates\cr
Any comparison of interested parameters can be formulated by
multiplying a contrast matrix C on LB and set \cr \cr
H0: C*LB=d for any vector of value d to be tested \cr \cr
We then have \cr \cr
      C*Lb~N(C*MU,C*SIGMA*C') \cr \cr
and \cr \cr
      (C*Lb-d)'*inv(C*SIGMA*C')*(C*Lb-d)~chisq(q,lambda) \cr \cr
where degree of freedom q=rank(C*SIGMA*C'), \cr
non-centrality parameter lambda=(C*LB-d)'*inv(C*SIGMA*C')*(C*LB-d) \cr \cr \cr
Power of the test H0 is given by 1-beta=P(chisq(q,lambda)>qchisq(1-alpha,lambda)) \cr
Required sample size for desired power can be obtained by bisection method.
}
\details{
\href{https://drive.google.com/open?id=1MQvzp9na-Zm9mFT5jC1I_3U0JHEFky_k}{Technical note}
}
\note{
License: GPL-3
}
\seealso{
\code{\link{lme.Lb.dist.theta}}, \code{\link{pass.lme.CLb.test}}
}
\author{
Marco Chak Yan YU \cr
Maintainer: Marco Chak Yan YU <marcocyyu@gmail.com>
}
